// music_app/app/artists/types.tsx
export interface Artist {
  idArtist: string;
  strArtist: string;
  strArtistStripped: string | null;
  strArtistAlternate: string;
  strLabel: string;
  idLabel: string;
  intFormedYear: string;
  intBornYear: string;
  intDiedYear: string | null;
  strDisbanded: string | null;
  strStyle: string;
  strGenre: string;
  strMood: string;
  strWebsite: string;
  strFacebook: string;
  strTwitter: string;
  strBiographyEN: string;
  strBiographyDE: string;
  strBiographyFR: string;
  strBiographyCN: string | null;
  strBiographyIT: string;
  strBiographyJP: string | null;
  strBiographyRU: string | null;
  strBiographyES: string;
  strBiographyPT: string | null;
  strBiographySE: string | null;
  strBiographyNL: string | null;
  strBiographyHU: string | null;
  strBiographyNO: string | null;
  strBiographyIL: string | null;
  strBiographyPL: string | null;
  strGender: string;
  intMembers: string;
  strCountry: string;
  strCountryCode: string;
  strArtistThumb: string;
  strArtistLogo: string;
  strArtistCutout: string;
  strArtistClearart: string;
  strArtistWideThumb: string;
  strArtistFanart: string;
  strArtistFanart2: string;
  strArtistFanart3: string;
  strArtistFanart4: string;
  strArtistBanner: string;
  strMusicBrainzID: string;
  strISNIcode: string | null;
  strLastFMChart: string;
  strLocked: string;
}